+++ lib/src/parsers/lcov.dart
@@ -8,16 +8,12 @@ import 'package:crypto/crypto.dart';
 import 'package:lcov/lcov.dart';
 import 'package:path/path.dart' as path;
 
-/// Parses the specified [LCOV](http://ltp.sourceforge.net/coverage/lcov.php) coverage [report].
+/// Parses the specified [LCOV](http://ltp.sourceforge.net/coverage/lcov.php) coverage report.
 Future<Job> parseReport(String report) async {
   var sourceFiles = <SourceFile>[];
   for (var record in new Report.fromCoverage(report).records) {
-    var source;
-    try { source = await new File(record.sourceFile).readAsString(); }
-    on Exception { throw new FileSystemException('Source file not found.', record.sourceFile); }
-
-    var lines = source.split(new RegExp(r'\r?\n'));
-    var coverage = new List<int>(lines.length);
+    var source = await new File(record.sourceFile).readAsString();
+    var coverage = new List<int>(source.split(new RegExp(r'\r?\n')).length);
     for (var lineData in record.lines.data) coverage[lineData.lineNumber - 1] = lineData.executionCount;
 
     var filename = path.relative(record.sourceFile);