+++ test/all.dart
@@ -4,12 +4,15 @@ import 'git_commit_test.dart' as git_commit_test;
 import 'git_data_test.dart' as git_data_test;
 import 'git_remote_test.dart' as git_remote_test;
 import 'job_test.dart' as job_test;
-import 'source_file_test.dart' as source_file_test;
-
+import 'parsers/clover_test.dart' as clover_test;
 import 'parsers/lcov_test.dart' as lcov_test;
+import 'source_file_test.dart' as source_file_test;
 
 /// Tests all the features of the package.
 void main() {
+  clover_test.main();
+  lcov_test.main();
+
   client_test.main();
   configuration_test.main();
   git_commit_test.main();
@@ -17,6 +20,4 @@ void main() {
   git_remote_test.main();
   job_test.main();
   source_file_test.main();
-
-  lcov_test.main();
 }